Move CollectOrderPaymentUseCaseError to Yosemite for reuse
