+++ WooCommerce/Classes/POS/Adaptors/Card Present Payments/CardPresentPaymentsAlertPresenterAdaptor.swift
@@ -2,6 +2,7 @@ import Foundation
 import Combine
 import enum Yosemite.ServerSidePaymentCaptureError
 import enum Yosemite.CardReaderServiceError
+import enum Yosemite.CollectOrderPaymentUseCaseError
 
 final class CardPresentPaymentsAlertPresenterAdaptor: CardPresentPaymentAlertsPresenting {
     typealias AlertDetails = CardPresentPaymentEventDetails
